+++ include/bleb/byteio_stdio.hpp
@@ -31,13 +31,13 @@ public:
         return ftell(file);
     }
 
-    virtual bool getBytesAt(uint64_t pos, uint8_t* buffer, uint64_t count) override {
+    virtual bool getBytesAt(uint64_t pos, uint8_t* buffer, size_t count) override {
         fseek(file, pos, SEEK_SET);
         assert(fread(buffer, 1, count, file) == count);
         return true;
     }
 
-    virtual bool setBytesAt(uint64_t pos, const uint8_t* buffer, uint64_t count) override {
+    virtual bool setBytesAt(uint64_t pos, const uint8_t* buffer, size_t count) override {
         fseek(file, pos, SEEK_SET);
         assert(fwrite(buffer, 1, count, file) == count);
         return true;
